add herencia to partidamultijudaor and partidainvididual classes (#17)

+++ Classes/Partida/Partida.cpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-Partida::Partida(DtFechaHora * fechaHora,int duracion){
+Partida::Partida(DtFechaHora * fechaHora,float duracion){
     this->fecha = fechaHora;
     this->duracion = duracion;
 }
